backend/pyproject.toml
[tool.poetry]
name = "face-attendance-backend"
version = "0.1.0"
description = "FastAPI backend for Face Recognition Attendance System"
authors = ["Your Name <you@example.com>"]
# Point to the repo-level README to satisfy Poetry packaging
readme = "../README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
pydantic = "^2.7.2"
pydantic-settings = "^2.4.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
email-validator = "^2.2.0"
fastapi-pagination = "^0.12.34"
pyjwt = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
httpx = "^0.27.0"
ruff = "^0.4.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

